Use of new installation manager class
